SG-6925 Build uncompressed uImage (uImage.bin)
Unfortunately, Yocto/OE has a bug, which causes u-boot-mkimage to be a
dependency for linux-yocto only when building *exactly* uImage, but not for
uImage.bin. Therefore, we work around this by just build uImage in addition.

When changing SERIAL_CONSOLES to "57600;ttyS2", the resulting uImage.bin can be
loaded via TFTP to my LinkIt Smart 7688 and actually boots and prints startup
messages via UART.

Of course it fails once the kernel figures out that there is no rootfs.